chore: Safer error handling
In a `catch`, the caught value is of type `unknown`.
This is because `throw "something unexpected happened"` is valid JS.
That is, you don't have to throw an `Error` type.

Refactor to handle this fact in a slightly more type-safe way.

See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/throw