fix: Remove unused code
We were previously reading these environment variables in order to manually fetch a token.

Since #17, we're using the functionality provided by `@actions/core` to do this,
so we no longer need to read these environment variables as we're not using them.

This is a `fix` because `getEnvString` will no longer throw unnecessarily.